Fix minor (undetectable) bug in src/diet.cc
Fix the creation of the matrix A to add opposite sign inequalities at
an offset of i + n as opposed to i + m, because m denotes the number of
unknowns in the optimization problem and n the number of equations.

Coincidentally, the judge tests are weak enough to rate the source 100%
with n and m interchangeable in this context.